# docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "OHLC is large amount of historical OHLC price data in CSV files format, which now needs to be centralized and digitized.",
        "title": "Historical OHCL Price Data docs",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
        "/data": {
            "get": {
                "description": "get",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "data"
                ],
                "summary": "returns all data.",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "page",
                        "name": "page",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 5,
                        "description": "per page",
                        "name": "per_page",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "symbol",
                        "name": "symbol",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "open",
                        "name": "open",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "unix",
                        "name": "unix",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "high",
                        "name": "high",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "low",
                        "name": "low",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "close",
                        "name": "close",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "created at",
                        "name": "created_at",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/outputForms.State"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/outputForms.State"
                        }
                    }
                }
            },
            "post": {
                "description": "post data",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "data"
                ],
                "summary": "post gives a files from formData and insert records quickly.",
                "parameters": [
                    {
                        "type": "file",
                        "description": "resource files",
                        "name": "files",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/outputForms.State"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/outputForms.State"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/outputForms.State"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "outputForms.Counts": {
            "type": "object",
            "properties": {
                "page": {
                    "type": "integer"
                },
                "per_page": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                },
                "total_page": {
                    "type": "integer"
                }
            }
        },
        "outputForms.State": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "counts": {
                    "$ref": "#/definitions/outputForms.Counts"
                },
                "data": {
                    "type": "any"
                },
                "detail": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "boolean"
                }
            }
        }
    }
}